# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LeetCode Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /push/subscribe:
    post:
      operationId: savePushSubscription
      tags: [push]
      summary: Subscribe to push notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavePushSubscriptionBody"
      responses:
        "200":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushSubscribeResponse"
        "401":
          description: Unauthorized
    delete:
      operationId: deletePushSubscription
      tags: [push]
      summary: Unsubscribe from push notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeletePushSubscriptionBody"
      responses:
        "200":
          description: Unsubscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushSubscribeResponse"
  /push/mock:
    post:
      operationId: sendMockNotification
      tags: [push]
      summary: Send a mock push notification for testing
      responses:
        "200":
          description: Mock notification sent
        "401":
          description: Unauthorized
  /profiles/{username}:
    get:
      operationId: getLeetcodeProfile
      tags: [profiles]
      summary: Get a LeetCode profile with recent solved problems and following list
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profile data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeetcodeProfile"
        "404":
          description: Profile not found
  /profiles/{username}/db-summary:
    get:
      operationId: getDbProfileSummary
      tags: [profiles]
      summary: DB-only profile summary
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Profile summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbProfileSummary"
        "404":
          description: Profile not found in DB
  /profiles/{username}/heatmap:
    get:
      operationId: getProfileHeatmap
      tags: [profiles]
      summary: Get daily solve counts for the last 365 days
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Heatmap data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HeatmapDataPoint"
  /profiles/{username}/save:
    post:
      operationId: saveProfileToDb
      tags: [profiles]
      summary: Fetch and save profile to DB (no follow)
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbProfileSummary"
components:
  schemas:
...
    SolvedProblem:
      type: object
      properties:
        id:
          type: integer
        leetcodeUsername:
          type: string
        problemSlug:
          type: string
        problemTitle:
          type: string
        difficulty:
          type: string
        solvedAt:
          type: string
        titleSlug:
          type: ["string", "null"]
        submissionId:
          type: ["string", "null"]
      required:
        - id
        - leetcodeUsername
        - problemSlug
        - problemTitle
        - difficulty
        - solvedAt
...
    LeetcodeFollowingEntry:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
      required:
        - username
    LeetcodeProfile:
      type: object
      properties:
        leetcodeUsername:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        totalSolved:
          type: ["integer", "null"]
        easySolved:
          type: ["integer", "null"]
        mediumSolved:
          type: ["integer", "null"]
        hardSolved:
          type: ["integer", "null"]
        recentProblems:
          type: array
          items:
            $ref: "#/components/schemas/SolvedProblem"
        following:
          type: array
          items:
            $ref: "#/components/schemas/LeetcodeFollowingEntry"
      required:
        - leetcodeUsername
        - recentProblems
        - following
    DbProfileSummary:
      type: object
      properties:
        leetcodeUsername:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        totalSolved:
          type: ["integer", "null"]
        solvedInPeriod:
          type: integer
        inDatabase:
          type: boolean
        recentSlugs:
          type: array
          items:
            type: string
      required:
        - leetcodeUsername
        - solvedInPeriod
        - inDatabase
...
    SavePushSubscriptionBody:
      type: object
      properties:
        endpoint:
          type: string
        p256dh:
          type: string
        auth:
          type: string
      required:
        - endpoint
        - p256dh
        - auth
    DeletePushSubscriptionBody:
      type: object
      properties:
        endpoint:
          type: string
      required:
        - endpoint
    HeatmapDataPoint:
      type: object
      properties:
        date:
          type: string
        count:
          type: integer
      required:
        - date
        - count
    PushSubscribeResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success